Create README.md
This Power BI report is created using the data gathered from a survey about the experiences of riding the S-Train and long-distance train lines. The report provides insights into the performance of the signaling systems and overall train punctuality. It includes visualizations that showcase the reliability and quality of onboard amenities and the demographic breakdown of the respondents. The report allows you to filter and drill down into the data, providing a comprehensive view of the survey results. Overall, this Power BI report is a valuable tool for analyzing the survey data and identifying areas for improvement in the train services.